+++ src/components/Row.js
@@ -328,8 +328,8 @@ const rowFunctionComponent = (props) => {
                      customProps={column.customProps}
                      customColumnClass={column.customColumnClass}
                      compressLongText={column.compressLongText}
-                     activateNonWorkingDaysFilter={props.activateNonWorkingDaysFilter}
                      tabIndex={props.getTabIndex ? props.getTabIndex(row, colIndex) : -1}
+                     filter_hermosillo_non_working_days={props.filter_hermosillo_non_working_days}
                   />
                )
             } else {