+++ src/components/Table/Table.js
@@ -1076,7 +1076,6 @@ class Table extends React.Component {
                      onDrop={this.props.onDropInRow}
                      scrollTo={this.props.scrollToRow}
                      handleFilterColumn={this.handleFilterColumn}
-                     activateNonWorkingDaysFilter={this.props.activateNonWorkingDaysFilter}
                      onContextMenu={(x, y, actions) => {
                         this.setState((prevState) => ({
                            contextMenu: {
@@ -1124,7 +1123,6 @@ class Table extends React.Component {
                type={this.props.type}
                customRowClass={this.props.customRowClass}
                ignoreItemStyle={this.props.ignoreItemStyle}
-               activateNonWorkingDaysFilter={this.props.activateNonWorkingDaysFilter}
             />
          )
       } else if (this.props.totalRowColumns) {
@@ -1148,7 +1146,6 @@ class Table extends React.Component {
                type={this.props.type}
                customRowClass={this.props.customRowClass}
                ignoreItemStyle={this.props.ignoreItemStyle}
-               activateNonWorkingDaysFilter={this.props.activateNonWorkingDaysFilter}
             />
          )
       }
@@ -1778,7 +1775,6 @@ class Table extends React.Component {
                            type={this.props.type}
                            customRowClass={this.props.customRowClass}
                            ignoreItemStyle={this.props.ignoreItemStyle}
-                           activateNonWorkingDaysFilter={this.props.activateNonWorkingDaysFilter}
                         />}
                      </thead>
                   </table>
@@ -2005,7 +2001,6 @@ Table.propTypes = {
    filterOptions: PropTypes.object,
    ignoreItemStyle: PropTypes.bool,
    shouldShowSelectIcon: PropTypes.bool,
-   activateNonWorkingDaysFilter: PropTypes.bool,
    allowTabNavigationForParents: PropTypes.bool,
    allowTabNavigationForChildren: PropTypes.bool,
 };